feat: moving room to report package
