+++ pkg/ws/report/room.go
@@ -1,4 +1,4 @@
-package common
+package report
 
 import (
 	"github.com/kptm-tools/core-service/pkg/domain"
